requirements.txt
Django==2.1.1
django-cors-headers==2.4.0
django-extensions==2.1.2
django-rest-framework==0.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
